Demonstrate argument parsing with `get_personalized_greeting`.
`get_personalized_greeting` can take a string (such as "Arthur"),
or no arguments at all. In the latter case, it returns the string
"hello whoever you are." It's equivalent to the Python code

    def get_personalized_greeting(who='whoever you are'):
        return "hello " + who

Unlike the very primitive `get_hello`, this one also raises the
appropriate exception types when it's called incorrectly.
And it reports a docstring appropriately, too.